Site rework
This site design is for a more functional and document driven site.
As part of this move we move from jekyll to hugo to make the development
process easy and quick and use the google/docksy theme which fits sites
like that, with documention of an opensource project in mind.

ADDING CONTENT
One of the main advantages is the ease of content generation from the
various markdown files in the parodos-dev repositories.

For example to create a documentation entry from the
serverless-workflows REAMDE.md in the root of the repo, create this
document

```
{{< remoteMD
"https://github.com/parodos-dev/serverless-workflows/blob/main/README.md?raw=true"
>}}

```

DEVELOPMENT
Local development requires git, go and hugo binary, which is easily
downloaded from hugo site. No need for speciel containers althought it
is just as easy.

To run the server locally:
```
hugo server
```

CONTENT ORGANIZATION
Being focused on documentation on development articles, the content goes
mainly under dedicated folders, in the form of markdown files:

    /content/docs
    /content/post

Based on that content the site will generate a sidebar with the
structure of all the files and folders.