Multi doc-version section
Motivation
Have a per-version documentation.

Modification
With this change the content/docs directory structure is changed
to content/$version/docs plus few other changes to the front-matter and
hugo.yaml to include the necessary changes to render a Documentation
drop-down menu.

The 'Blog' section remains the same, as blogs are have their own pace
and are not meant to be fixed per version, blogs are articles, not
documentation.

Action items for the team: version 1.1.x now has a folder with a copy of
the main documentation and it needs to be updated accordingly. Also some
of the changes in 'main' may be 1.1.x specific and may need a change
just as well.

Signed-off-by: Roy Golan <rgolan@redhat.com>