Copy referenced content to fix 1.2 documentation
Issue: The 1.2 documentation is referencing "main" branches of external projects.

Unless we can branch-out those projects for 1.2, we need to make a copy
of the content to lock the changes. Otherwise, ongoing development would
make this 1.2 documentation unmatching with the 1.2 features.

The "main" documentation is untouched - keeps referencing those external document.
When time of 1.3 comes, we can either copy actual referenced content or
branch-out those external projects and update links in this project.